moved renderer header into renderer folder also made a new renderer clear that takes in colors in rgb 0-1 floats and has a check that checks for values outside this range and if there are any warns the user
