+++ opengl tower defence game/imgui.ini
@@ -1,5 +1,5 @@
 [Window][Debug##Default]
-Pos=1213,310
+Pos=526,216
 Size=615,430
 Collapsed=0
