+++ opengl tower defence game/src/Renderer/Shader.cpp
@@ -1,6 +1,6 @@
 #include "Shader.h"
 
-#include "../Renderer.h"
+#include "Renderer.h"
 
 #include <iostream>
 #include <fstream>
