+++ opengl tower defence game/src/Renderer/Texture.h
@@ -1,6 +1,5 @@
 #pragma once
-
-#include "../Renderer.h"
+#include <string>
 
 class Texture
 {
