+++ opengl tower defence game/src/Renderer/VertexArray.cpp
@@ -1,5 +1,5 @@
 #include "VertexArray.h"
-#include "../Renderer.h"
+#include "Renderer.h"
 #include "VertexBufferLayout.h"
 
 VertexArray::VertexArray()
