+++ opengl tower defence game/src/Renderer/VertexBufferLayout.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <vector>
-#include "../Renderer.h"
 #include <GL/glew.h>
 #include <stdexcept>
