+++ opengl tower defence game/src/Tests/TestClearColor.cpp
@@ -1,10 +1,10 @@
 #include "TestClearColor.h"
-#include "../Renderer.h"
+#include "../Renderer/Renderer.h"
 #include "../Vendor/imgui/imgui.h"
 
-namespace test 
+namespace test
 {
-	
+
 	TestClearColor::TestClearColor()
 		: m_ClearColor { 0.2f, 0.3f, 0.8f, 1.0f }
 	{
